backend/requirements.txt
flask
flask-cors
flask-bcrypt
flask-sqlalchemy
python-dotenv
flask-jwt-extended